requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==2.3.7
PyMySQL==1.1.0
SQLAlchemy==2.0.20
gunicorn==21.2.0
python-dotenv==1.0.0